유저 식물 등록 일부 구조 분해
