+++ frontend/src/components/userplant/PlantObserveForm.jsx
@@ -1,0 +1,5 @@
+const PlantObserveForm = () => {
+  return <div className="text-gray-600 text-sm">일지 작성은 준비중입니다.</div>;
+};
+
+export default PlantObserveForm;
